Move graph_construct integration test into own file
This feature is getting more complex so lets have another file instead of a very large one
